// ionic/RVSHealthyIndia/src/models/child/child.model.ts
export interface Child {
    key?: string;
    asha_worker_id: string;
    birth_certificate_number: string;
    child_id: string;
    child_name: string;
    child_blood_group: string;
    dob: string;
    gender: string;
    height: string;
    weight: string;
    husband_name: string;
    mother_blood_group: string;
    mother_id: string;
    mother_name: string;
    phone_number_one: string;
    phone_number_two: string;
}